cl: get cl tests building again
