+++ cl/cl/Cargo.toml
@@ -12,8 +12,8 @@ risc0-groth16 = "1.0.1"
 blake2 = "0.10.6"
 # jubjub = "0.10.0"
 group = "0.13.0"
+rand = "0.8.5"
 rand_core = "0.6.0"
-rand_chacha = "0.3.1"
 lazy_static = "1.4.0"
 hex = "0.4.3"
 curve25519-dalek = {version = "4.1", features = ["serde", "digest", "rand_core"]}
