--- cl/cl/src/test_util.rs
@@ -1,7 +1,0 @@
-use rand_core::SeedableRng;
-
-pub fn seed_rng(seed: u64) -> impl rand_core::RngCore {
-    let mut bytes = [0u8; 32];
-    bytes[..8].copy_from_slice(&seed.to_le_bytes());
-    rand_chacha::ChaCha12Rng::from_seed(bytes)
-}
